+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+**v2.4.0** Extend LibraryProxy to include mouseEvent and keyEvent.
+
+**v2.3.4** Simplify control_panel library (replacing `c.title = 'PaneTitle'` with `c.title('PaneTitle')`) also enable use of `block` with `button's`.
+
+**v2.3.3** Update to processing-3.3.4, and upgrade jruby-9.1.12.0 last in 9.1 series?
+
+**v2.3.3** Update to processing-3.3.2
+
 **v2.3.1** Update to processing-3.3.1, Update samples which now include `#!/usr/bin/env jruby` to assist running sketches from atom, add Rakefile for automatic sketch running for many samples, expected upgrade to final jruby-9.1.x.x
 
 **v2.3.0** Update to core processing-3.3 `grid` method now implemented in java, bump examples to include WOVNS